backend_deploy/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-docx
pypdf
pandas
openpyxl
python-jose[cryptography]
passlib[bcrypt]
firebase-admin
google-generativeai
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
langchain
langchain-openai
langchain-community
python-multipart
gunicorn
httpx
beautifulsoup4
lxml
